requirements.txt
flask
gunicorn
flask_cors
mysql-connector-python
pandas
numpy
scikit-learn
pymongo
pymongo[srv]